add SpicyBurger enum
